[unorganized] adjustments made for pi.
This commit consists of unorganized adjustments made while the code was loaded on the pi. They mostly consist of various quality of life improvements--the most significant is the addition of a command-line utility for sending commands to the arduino and the configuration files I refactored out of the main code itself to allow the command line utility to run off of the same information.